Funções para a avaliação do dentista implementadas, lista de avaliações feita.

+++ app/src/main/java/br/edu/puccampinas/pi3/Avaliacoes.kt
@@ -1,3 +1,3 @@
 package br.edu.puccampinas.pi3
 
-data class Avaliacoes(val nome: String, val estrela: Int, val comentario: String)
+data class Avaliacoes(val nome: String, val estrela: Long, val comentario: String)
